+++ src/pages/lamp/FirstthemeWritePage.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import styled from "styled-components";
 import { useRecoilState } from "recoil";
+import { useNavigate } from "react-router-dom";
 import {
   nicknameState,
   wishState,
@@ -23,11 +24,12 @@ const Background = styled.div`
   background-size: cover;
 `;
 
-const SecondthemeWrite = () => {
+const FirstthemeWritePage = () => {
   const [nickname, setNickname] = useRecoilState(nicknameState);
   const [wish, setWish] = useRecoilState(wishState);
   const [email, setEmail] = useRecoilState(emailState); // 이메일 상태 추가
   const { backgroundImageUrl } = GradientBackground();
+  const navigate = useNavigate();
 
   // 모든 입력란이 채워졌는지 확인하는 함수
   const isSatisfied = () => {
@@ -40,8 +42,8 @@ const SecondthemeWrite = () => {
         const postData = {
           nickname: nickname,
           content: wish,
-          email: email, // 질문
-          theme: 2
+          email: email,
+          theme: 1
         };
 
         // 백엔드 데이터 전송
@@ -49,7 +51,10 @@ const SecondthemeWrite = () => {
 
         // POST 성공 시
         console.log("등불이 저장되었습니다", response.data);
-        alert("소원이 이뤄질 거예요");
+        // alert("소원이 이뤄질 거예요");
+
+        // Navigate to /flyLamp
+        navigate("/flyLamp");
       } catch (error) {
         console.error("POST ERROR", error);
         alert("Failed to submit data. Please try again later.");
@@ -62,7 +67,7 @@ const SecondthemeWrite = () => {
   return (
     <Background $backgroundImageUrl={backgroundImageUrl}>
       <Header title="등불 작성하기" />
-      <DescriptionText preText="올해가 지나고 난 어떤 것을 이뤘을까?" />
+      <DescriptionText preText="이번 학기가 끝났을 때 나는 어떤 모습일까?" />
       <LanternWishPaper inputType="email" />
       <LanternWritebutton
         onClick={nextBtnOnClick}
@@ -73,4 +78,4 @@ const SecondthemeWrite = () => {
   );
 };
 
-export default SecondthemeWrite;
+export default FirstthemeWritePage;
